// src/shared/enums/xikasu.source.enum.ts
export enum XiKasuSourceEnum {
    XiFiFree = "XiFiFree",
    Voucher = "Voucher",
    XiFiNow = "XiFiNow",
    XiFiMax = "XiFiMax",
    XiFiConnect = "XiFiConnect",
    XiKasuAd = "XiKasuAd",
    ReferrerBonus = "ReferrerBonus",
    RefereeBonus = "RefereeBonus",
    Payment = "Payment",
    RefundPayment = "RefundPayment"
}